+++ soroban-sdk/src/deployer.rs
@@ -1,14 +1,6 @@
 use crate::{
-    env::internal::Env as _, namespace::DeployerNamespace, Bytes, BytesN, Env, TryFromVal,
+    env::internal::Env as _, namespace::{DeployerNamespace, CurrentNamespace}, Bytes, BytesN, Env, TryFromVal,
 };
-
-pub trait DeployWasm {
-    fn deploy_wasm(&self, env: &Env, salt: Bytes, wasm: Bytes) -> BytesN<32>;
-}
-
-pub trait DeployToken {
-    fn deploy_token(&self, env: &Env, salt: Bytes) -> BytesN<32>;
-}
 
 pub struct Deployer {
     pub env: Env,
@@ -38,5 +30,27 @@ impl Deployer {
                 todo!()
             }
         }
+    }
+}
+
+trait DeployWasm {
+    fn deploy_wasm(&self, env: &Env, salt: Bytes, wasm: Bytes) -> BytesN<32>;
+}
+
+trait DeployToken {
+    fn deploy_token(&self, env: &Env, salt: Bytes) -> BytesN<32>;
+}
+
+impl DeployWasm for CurrentNamespace {
+    fn deploy_wasm(&self, env: &Env, salt: Bytes, wasm: Bytes) -> BytesN<32> {
+        let id = env.create_contract_from_contract(wasm.to_object(), salt.to_object());
+        BytesN::<32>::try_from_val(env, id).unwrap()
+    }
+}
+
+impl DeployToken for CurrentNamespace {
+    fn deploy_token(&self, env: &Env, salt: Bytes) -> BytesN<32> {
+        let id = env.create_token_from_contract(salt.to_object());
+        BytesN::<32>::try_from_val(env, id).unwrap()
     }
 }
